+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+# 1.1.0
+
+- update for Unity 2022.3.22f1
+- update for VRC SDK 3.7.1 (Sep 2024)
+- fix uploading disabled avatars
+- fix error when SDK hasn't been opened yet
+- improved reliability of statuses
+- added buttons to build or build and test all avatars
+- added reset statuses button
+- added user friendly logs
+
 # 1.0.0
 
 - first release